pyproject.toml
[tool.poetry]
name = "subliminal"
version = "0.1.0"
description = ""
authors = ["ouzb64ty <theoph.po@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyseagull = "^1.0.0-beta.4"
pytest = "^6.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
